Beginning some code cleanup/refactoring, still in progress...
